Add a note about generating config file to Laravel Mix section
Hello, 

I was doing a fresh install with Laravel and was a bit confused how to generate the configuration file, didn't take long to find the right 
section. But could be useful to point that out inline with the `installation` section. 

No worries if you decide not necessary! Have a great week!

V/R, 

Wyatt