Update installation.blade.md
`tailwind` refers to a different package in the npm registry. It is dangerous to attempt to run `npx tailwind` without prior explicit warning that people should install `tailwindcss` first. The general public uses npx to execute npm packages without having to install them locally, therefor running `npx tailwind init` results in unexpected behaviour. The proof is that there already are issues on the web of users failing to run it.

Using this alias is rather unfortunate and I see three possible solutions:
1) update the commands to not attempt to install a package if it's not available locally `--no-install` as in this PR
2) warn people explicitly that `tailwind` is just an alias and that `tailwindcss` must be installed first
3) use `tailwindcss` alias for the executable instead of `tailwind`