Add focus:outline-none to example code
See https://github.com/tailwindlabs/tailwindcss/issues/4204 (Adam said that he already updated the docs, but I don't see it here. Maybe it was in another place.)

When I first tried the example code, I didn't understand why the focus ring was still showing when I clicked the button with the mouse. I thought that using `focus-visible` alone would only show the ring when using the keyboard. It's only when expecting the source code that I understood that `focus:outline-none` was necessary (which on a second thought is actually quite logical).