Add ts,tsx to guides/create-react-app.js
I was stuck on an issue for about an hour where my typescript files weren't getting styles and this was the culprit. It's a little obvious and I should have seen it coming, but mentioning it in the docs might help the next person using typescript and trying to add tailwind.