Add note on exclusion glob patterns
As someone who is not familiar with glob patterns, I just spent a good chunk of time trying to figure out if excluding patterns was possible. It doesn't mention it anywhere in the tailwind documentation, so I thought it would be a good idea to add it here as a small note for posterity.