Correct Fresh Prince lyrics
Could there be anything more vital to this project than having the Fresh Prince lyrics correct? I don't think so. 

This proposed change alters the line "And I like to take a minute and sit right here" to "And I'd like to take a minute just sit right there". To validate:

* https://genius.com/Dj-jazzy-jeff-and-the-fresh-prince-the-fresh-prince-of-bel-air-lyrics
* https://www.songlyrics.com/pellek/yo!-home-to-bel-air-the-fresh-prince-of-bel-air-opening-theme-lyrics/
* https://freshprince.fandom.com/wiki/Yo_Home_To_Bel-Air