Remove inaccruate statements on inline styles
The moment CSS variables arrived on the scene, the `style=` attribute gained superpowers. The information currently on the Tailwind website is ~4 years out of date.

> Using inline styles, every value is a magic number

Not true. If you want, every value you put in an inline style can be a CSS variable and utilize a full design system. Not magic at all.

> You can't use media queries in inline styles

No, but the CSS variables you use can themselves be redefined by media queries. And you can also build things like fluid sizes and fluid typography out of CSS variables.

> Inline styles can't target states like hover or focus

Again, not true. With CSS variables you can redefine the value of the variable based on any state possible in a stylesheet, including hover or focus states.

This PR removes the inaccurate statements. I'm not sure what you would replace them with, but hopefully it would not contain information ~4 years out of date.