Provide an example of custom size
The documentation shows using curly bracket for the size portion which I tried to do and it does not work you need a bracket instead. This might be explained else were in the documentation but it would be good to have at least one example on the page that uses a custom size everywhere on the website that `-{size}` is mentioned.