Fix `create astro` link (#2056)
The first step of the [Astro framework
guide](https://tailwindcss.com/docs/installation/framework-guides/astro)
links to the Astro documentation for `create astro` using a link to the
docs landing page and an anchor that doesn't exist on this page.

After discussing with the Astro docs team, this PR updates the link to
point to
https://docs.astro.build/en/install-and-setup/#install-from-the-cli-wizard
which is a more appropriate link for the `create astro` command.