Clarify usage of outline utilities in documentation (#2071)
This pull request includes a small change to the
`src/docs/outline-width.mdx` file. The change modifies the description
to clarify the usage of `outline` utilities.

*
[`src/docs/outline-width.mdx`](diffhunk://#diff-a7a0467902b75954570708cc09d471f346011c3bce60b92e96aaa0c8b21f56daL22-R22):
Changed "and" to "or" in the description to clarify that either
`outline` or `outline-<number>` utilities can be used to set the width
of an element's outline.

---------

Co-authored-by: Philipp Spiess <hello@philippspiess.com>