Fix color flash on page load when setting dark/light mode
Previously, when loading the landing page, the page colors would flash
briefly. For light mode the background flashes black before the white
background loads.

For dark mode, the header would flash white, and the text black, before
the correct colors render.

This was happening because the script was being executed after the page
had finished loading, likely due to how the page is packaged into
chunks.

To fix this, the dark mode script is now set inline, so will be part of
the generated HTML.

This fixes https://github.com/tailwindlabs/tailwindcss.com/issues/2105