Improve the `@custom-variant` docs (#2182)
This PR improves the `@custom-variant` docs, including:

1. Updating the "Adding custom variants" section on the "Adding custom
styles" page to actually explain the `@custom-variant` directive,
instead of the `@variant` directive.
2. Creating a new "Using variants" section on the "Adding custom styles"
page to explain the `@variant` directive.
3. Updating the "@custom-variant" section of the "Functions and
directives" page to remove the "pointer-coarse" example (in preparation
for v4.1), and also link to the new "Adding custom variants" section on
the "Adding custom styles" page, instead of the "Registering a custom
variant" on the "Hover, focus, and other states" page. The "Adding
custom styles" page is now considered the canonical location for this
information.
4. Simplifying the "Registering a custom variant" section on the "Hover,
focus, and other states" page, and adding a link to the "Adding custom
variants" section on the "Adding custom styles" page for more
information.