Added anchors to custom headings (#2241)
Fixes: #2240 

In the spirit of reusability, I declared a new component called
`Heading`. You need to specify the level (between 1 and 6), the ID
(which is mandatory), and the content of the element. With this in
place, I also declared `H2` through `H6` elements. Thus, instead of:

```html
<h3 id="import-directive">@import</h3>
```
*
[Preview](https://tailwindcss.com/docs/functions-and-directives#import-directive)

you can use:
```html
<H3 id="import-directive">@import</H3>
```
*
[Preview](https://tailwindcss-com-git-fork-rozsazoltan-fix-he-f001db-tailwindlabs.vercel.app/docs/functions-and-directives#import-directive)

what will be:
```html
<h3 id="import-directive">
  <a href="#import-directive" class="anchor">@import</a>
</h3>
```

---------

Co-authored-by: Philipp Spiess <hello@philippspiess.com>