Fix word highlights (#2250)
This basically reverts the FF fix from #2126 which fixed FF but broke
word highlights. Also divs aren't allowed in `<code>` elements so
Shiki's assumption that it really only needs to look inside `<span>`s
isn't a terrible one in this case.

We'll have to come up with some other solution to work around the FF bug