Add details about `hidden` attribute preflight rules (#2321)
In v4 we made a change that makes elements with a hidden attribute
`display: none !important`. A consequence of this is that `<div hidden
class="block">test</div>` will be invisible in v4 where it was visible
in v3.

The intent is that you must *remove* this attribute if the element isn't
actually meant to be hidden but we didn't document this in the upgrade
guide or the preflight page.

Closes https://github.com/tailwindlabs/tailwindcss/issues/18653

---------

Co-authored-by: Philipp Spiess <hello@philippspiess.com>
Co-authored-by: Nicolas <nicolas.walger@gmail.com>