Fix `text-pretty` description (#2210)
Current `text-pretty` description says,

> Use the `text-pretty` utility to prevent orphans (a single word on its
own line) at the end of a text block:

See [this recent blog post from Webkit about `text-wrap:
pretty;`](https://webkit.org/blog/16547/better-typography-with-text-wrap-pretty/)
and in particular,

> Because of the way Chrome’s implementation of pretty has been taught,
a lot of web developers expect this value is only supposed to prevent
short last lines. But that was never the intention. In fact, the CSS
Working Group defined a different value for such a purpose. It was [just
renamed](https://github.com/w3c/csswg-drafts/issues/11283) last week to
`text-wrap: avoid-short-last-lines`.

So it seems an updated description is needed and this PR has a first
attempt.

---------

Co-authored-by: Jordan Pittman <jordan@cryptica.me>
Co-authored-by: Philipp Spiess <hello@philippspiess.com>