Improve `place-content` wording by referencing the correct axis (#2358)
This PR fixes an issue where we were only using `block axis` in the
`place-content` property docs, but the `place-content` property works on
both the inline axis and the block axis.

This PR now references both the inline axis and block axis. This now
also talks about the difference between rows and columns because we only
talked about rows before.

Fixes: https://github.com/tailwindlabs/tailwindcss.com/issues/2357

---------

Co-authored-by: Jordan Pittman <jordan@cryptica.me>
Co-authored-by: Jordan Pittman <thecrypticace@gmail.com>