Fix incorrect CSS config path example in Nuxt 4 framework guide
In #2309, the Nuxt framework guide was updated to be compatible with
Nuxt 4. The `css` config path was updated in the instructions on this
page, but not also in the code example. This change updates the code
example to also use the new path compatible with Nuxt 4
("./app/assets/css/main.css" instead of "~/assets/css/main.css").