Clarify some details around `@reference` (#2376)
Closes #2374 

I opted to also add a section on subpath imports to note that they're
usable with `@import`, `@reference`, `@plugin`, and `@config`.

---

Closes #2377

I've added a note to the upgrade guide as well about arbitrary values in
grid-cols, grid-rows, and object-position — while this behavior was
related to v2 compat it was still supported in v3 so it feels fine to
call it out as having changed.